RUB-1127: carry exact u128 fee and sum_fees through both clients
Widen the authoritative transaction fee and block sum_fees from u64 to an
exact u128 scalar across consensus derivation, block aggregation, the
coinbase bound, node policy admission/ordering/relay, and both consensus
CLIs. Per-output and per-UTXO values stay u64; only the per-transaction
sums, their difference, and the block aggregate widen.

A fee above u64 is now accepted whenever ordinary validation succeeds:
width alone never yields TX_ERR_PARSE or TX_ERR_VALUE_CONSERVATION, and
sum_fees crossing u64 no longer yields BLOCK_ERR_PARSE. Fee-rate
comparisons cross-multiply over all 192 product bits of a u128 fee by a
u64 weight, so no comparison truncates, saturates, or uses floating point.

Widened JSON values are written as canonical unsigned decimal strings and
read from either that form or a legacy numeric token bounded to u64.

+++ clients/rust/crates/rubin-consensus/src/block_basic.rs
@@ -162,7 +162,7 @@ pub fn validate_block_basic_with_context_and_fees_at_height(
     block_height: u64,
     prev_timestamps: Option<&[u64]>,
     already_generated: u128,
-    sum_fees: u64,
+    sum_fees: u128,
 ) -> Result<BlockBasicSummary, TxError> {
     validate_block_basic_with_context_and_fees_at_height_and_rotation(
         block_bytes,
@@ -186,7 +186,7 @@ pub fn validate_block_basic_with_context_and_fees_at_height_and_rotation(
     block_height: u64,
     prev_timestamps: Option<&[u64]>,
     already_generated: u128,
-    sum_fees: u64,
+    sum_fees: u128,
     rotation: Option<&dyn RotationProvider>,
 ) -> Result<BlockBasicSummary, TxError> {
     // G.9: parse once, share `pb` between basic validation and the
